[FEATURE] Cronjob added and small restructure
- Moved logic of adding all translations to the database to a new model TranslationCollector.
- Added crontab to enrich database with new entries every night at 3:13 AM (to avoid conflict with other cronjobs).
- Added translation files for the module itself
- Removed delete option to avoid mistakes by merchants/clients
