--- Ui/Component/Listing/Column/Different.php
@@ -1,49 +1,0 @@
-<?php
-/**
- * Copyright © 2016 Magento. All rights reserved.
- * See COPYING.txt for license details.
- */
-namespace Experius\MissingTranslations\Ui\Component\Listing\Column;
-
-/**
- * Class Different
- */
-class Different extends \Magento\Ui\Component\Listing\Columns\Column
-{
-    /**
-     * @param \Magento\Framework\View\Element\UiComponent\ContextInterface $context
-     * @param \Magento\Framework\View\Element\UiComponentFactory $uiComponentFactory
-     * @param \Magento\Framework\UrlInterface $urlBuilder
-     * @param array $components
-     * @param array $data
-     */
-    public function __construct(
-        \Magento\Framework\View\Element\UiComponent\ContextInterface $context,
-        \Magento\Framework\View\Element\UiComponentFactory $uiComponentFactory,
-        \Magento\Framework\UrlInterface $urlBuilder,
-        array $components = [],
-        array $data = []
-    ) {
-        $this->urlBuilder = $urlBuilder;
-        parent::__construct($context, $uiComponentFactory, $components, $data);
-    }
-
-    /**
-     * Prepare Data Source
-     *
-     * @param array $dataSource
-     * @return array
-     */
-    public function prepareDataSource(array $dataSource)
-    {
-        if (empty($dataSource['data']['items'])) {
-            return $dataSource;
-        }
-
-        foreach ($dataSource['data']['items'] as & $item) {
-            $item['different'] = ($item['string'] !== $item['translate']) ? '1' : '0';
-        }
-
-        return $dataSource;
-    }
-}
