+++ etc/crontab.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Cron:etc/crontab.xsd">
+    <group id="default">
+        <job name="update_all_translations_to_database" instance="Experius\MissingTranslations\Cron\Collect" method="execute">
+            <schedule>13 3 * * *</schedule>
+        </job>
+    </group>
+</config>
