+++ i18n/en_US.csv
@@ -1,0 +1,1 @@
+"Has a translation", "Has a translation"
