+++ i18n/nl_NL.csv
@@ -1,0 +1,1 @@
+"Has a translation", "Heeft een vertaling"
